profiles/amd64: Drop mask on cryptograph add mask on pycparser-2.19
Issues with cryptography are related to pycparser-2.19 and/or
cffi-1.12.1, but prsently seems to be the former. Thus adding mask on
that package, and dropping mask on cryptography. It also effects other
packages like bcrypt, zstandard, and others. All same issue as reported
about cryptography in the following upstream bug.

https://bugs.gentoo.org/671948