dev-java/javamail: Revert 1.6.4-> 1.6.3, eclipse-ee4j/mail/issues/395
Upstream, months later, decides to replace JavaMail 1.6.4 tag with
JakartaMail 1.6.4 tag, with the sources being named 1.6.4.{zip,tar.gz}.
The new 1.6.4 source archives do not contain actual sources, but contain
something else that is usless. The package is messed up all around as part
of botched renaming. Eclipse-EE4J is back to removing and replacing tags,
months after those tags were created. That horrible pratice created a huge
mess in issue #67. It is unfortunate that it is occuring again.
Thankfully, it is just a single package this time. Clearly, upstream needs
to work on their release and re-tagging policies...

Deleteing tags is BAD, a horrible practice!!!
Replacing a deleted tag with the exact same version is even worse!!!

Is this really how enterprise software is developed and released?

https://github.com/Obsidian-StudiosInc/os-xtoo/issues/67
https://github.com/eclipse-ee4j/mail/issues/395

Package-Manager: Portage-2.3.67, Repoman-2.3.16