profiles/workstation: Add server USE to dev-db/firebird
Stupid change of USE flag from client to server, for a server package...
Renamed by upstream, to negate my use of client USE flag. By default,
the package installs a server, and it should OPTIONALLY install only client
libraries. This reversal is extremely odd, and it seems to be done out of
ill will.